reorganize data and model downloads
